Allow users to use "zebrium.com/ze_logtype" label/annotation to set logbasename
By default logbasename is the same as container name. However, sometimes
the container name is not right, for example, it can be <app_name>-<id> or
<app_name>-<timestamp>. With this change we allow users to set logbasename
with either label or annotation.

Please note on kubernetes label/annotation is applied on pod. So this method
should not be used on pod with different types of containers.